Add deduplication to infinite feed scroll example
This change introduces a simple deduplication mechanicsm for the example infinite
scroll feed code we provide for the agent. I have also added a 'isFetchingNextPage'
check to avoid the potential scenario of aggressively overfetching via quick scrolling.

If a post exists in the feed already (it has been 'seen'), the it and all duplicate
post events will be filtered out from the returned content.